% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_median_nltts.R
\name{calc_median_nltts}
\alias{calc_median_nltts}
\title{For each input file, calculate the median nLTT statistic.
Put these, with the parameters in a data frame}
\usage{
calc_median_nltts(folder_name, output_filename)
}
\arguments{
\item{folder_name}{name of the folder that contains data files,
as created by \code{\link{create_output_file}}}

\item{output_filename}{name of a comma-seperated file
containing parameters and median nLTTs}
}
\description{
For each input file, calculate the median nLTT statistic.
Put these, with the parameters in a data frame
}
\author{
Richel J.C. Bilderbeek
}
