% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_raket_params.R
\name{check_raket_params}
\alias{check_raket_params}
\title{Check if the \code{raket_params} are valid raket parameters.}
\usage{
check_raket_params(raket_params)
}
\arguments{
\item{raket_params}{\code{raket} parameters for one experiment,
as can be created by \code{\link{create_raket_params}}}
}
\description{
Will \link{stop} if the \code{raket_params} are invalid,
else will do nothing
}
\author{
Richel J.C. Bilderbeek
}
