% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_raket_params_filenames.R
\name{check_raket_params_filenames}
\alias{check_raket_params_filenames}
\title{Check is a \code{raket_params} hold Peregrine-friendly filenames
that follow the \code{raket} naming convention.}
\usage{
check_raket_params_filenames(raket_params)
}
\arguments{
\item{raket_params}{\code{raket} parameters for one experiment,
as can be created by \code{\link{create_raket_params}}}
}
\description{
Check is a \code{raket_params} hold Peregrine-friendly filenames
that follow the \code{raket} naming convention.
}
