% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_raket_params_list_elements.R
\name{check_raket_params_list_elements}
\alias{check_raket_params_list_elements}
\title{Check if the \code{raket_params} is a list with all the required elements}
\usage{
check_raket_params_list_elements(raket_params)
}
\arguments{
\item{raket_params}{\code{raket} parameters for one experiment,
as can be created by \code{\link{create_raket_params}}}
}
\description{
Check if the \code{raket_params} is a list with all the required elements
}
