% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_raket_paramses.R
\name{check_raket_paramses}
\alias{check_raket_paramses}
\title{Check if the argument is a list of \code{raket_params}}
\usage{
check_raket_paramses(raket_paramses)
}
\arguments{
\item{raket_paramses}{a list of \code{raket_params},
as created by \code{\link{create_general_params_set}}
or \code{\link{create_sampling_params_set}}}
}
\description{
Check if the argument is a list of \code{raket_params}
}
\author{
Richel J.C. Bilderbeek
}
