% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sampling_method.R
\name{check_sampling_method}
\alias{check_sampling_method}
\title{Checks if the \code{sampling_method} is a valid
sampling method. See \link{rkt_get_sampling_methods}
for a vector of valid sampling methods}
\usage{
check_sampling_method(sampling_method)
}
\arguments{
\item{sampling_method}{method how the incipient species representing a
species is picked. Must be in \link{rkt_get_sampling_methods}.}
}
\description{
Checks if the \code{sampling_method} is a valid
sampling method. See \link{rkt_get_sampling_methods}
for a vector of valid sampling methods
}
