% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_pbd_params.R
\name{collect_pbd_params}
\alias{collect_pbd_params}
\title{Collect the PBD parameters from a \code{raket} experiment}
\usage{
collect_pbd_params(project_folder_name)
}
\arguments{
\item{project_folder_name}{name of project folder}
}
\value{
a data frame the folder names and PBD parameters
}
\description{
Collect the PBD parameters from a \code{raket} experiment
}
\author{
Richel J.C. Bilderbeek
}
