% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkt_plot.R
\name{create_fig_1}
\alias{create_fig_1}
\title{Plot violin plot}
\usage{
create_fig_1(df_long = utils::read.csv("result_long.csv"),
  pdf_filename = tempfile(fileext = ".pdf"))
}
\arguments{
\item{df_long}{data frame in the long form}

\item{pdf_filename}{name of the PDF with the plot}
}
\description{
Plot violin plot
}
