% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_general_params_set.R
\name{create_general_params_set}
\alias{create_general_params_set}
\title{Creates the parameter set for the general exploration}
\usage{
create_general_params_set(
  project_folder_name = file.path(peregrine::get_pff_tempdir(), "raket_werper"),
  n_replicates = 1
)
}
\arguments{
\item{project_folder_name}{name of project folder}

\item{n_replicates}{number of replicates per biological parameter set}
}
\value{
a list of which each element is a \code{raket_params}
}
\description{
Creates the parameter set for the general exploration
}
\author{
Richel J.C. Bilderbeek
}
