% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_general_pbd_paramses.R
\name{create_general_pbd_paramses}
\alias{create_general_pbd_paramses}
\title{Create the general PBD parameter sets}
\usage{
create_general_pbd_paramses()
}
\value{
a list of \code{pbd_params}
}
\description{
Create the general PBD parameter sets
}
