% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_input_files.R
\name{create_input_files}
\alias{create_input_files}
\title{Creates the parameter files from the \code{raket_params}}
\usage{
create_input_files(raket_paramses)
}
\arguments{
\item{raket_paramses}{a list of \code{raket_params},
as created by \code{\link{create_general_params_set}}
or \code{\link{create_sampling_params_set}}}
}
\value{
The filenames of all parameter files created
}
\description{
Creates the parameter files from the \code{raket_params}
}
\seealso{
Use \link{create_input_files_test} to create the parameters file
for a test input set.
Use \link{create_input_files_general} to create the parameters file
for a general input set.
Use \link{create_input_files_sampling} to create the parameters file
for a sampling input set.
}
\author{
Richel Bilderbeek
}
