% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_input_files.R
\name{create_input_files_sampling}
\alias{create_input_files_sampling}
\title{Creates the parameter files in the article for the general exploration}
\usage{

  create_input_files_sampling(raket_paramses = create_sampling_params_set(project_folder_name
  = getwd()))
}
\arguments{
\item{raket_paramses}{a list of \code{raket_params},
as created by \code{\link{create_general_params_set}}
or \code{\link{create_sampling_params_set}}}
}
\value{
The filenames of all parameter files created
}
\description{
Creates the parameter files in the article for the general exploration
}
\author{
Richel Bilderbeek
}
