% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_input_files.R
\name{create_input_files_test}
\alias{create_input_files_test}
\title{Creates the parameter files in the article for the general exploration}
\usage{
create_input_files_test(test_params_set = create_test_params_set(),
  folder_name = getwd())
}
\arguments{
\item{folder_name}{name of the folder where all files are created}

\item{sampling_params_set}{the set of parameters,
as created by \code{\link{create_sampling_params_set}}}
}
\value{
The filenames of all parameter files created
}
\description{
Creates the parameter files in the article for the general exploration
}
\author{
Richel Bilderbeek
}
