% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_nltt_file.R
\name{create_nltt_file}
\alias{create_nltt_file}
\title{Measure the nLTT statistics from a BEAST2 run}
\usage{
create_nltt_file(input_filename, output_filename)
}
\arguments{
\item{input_filename}{input filename, 
containing the BEAST2 run results, 
as created by create_output_file}

\item{output_filename}{output filename, 
containing the nLTT values}
}
\description{
Measure the nLTT statistics from a BEAST2 run
}
