% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_output_file.R
\name{create_output_file}
\alias{create_output_file}
\title{Process an input file}
\usage{
create_output_file(input_filename = input_filename,
  output_filename = output_filename, verbose = FALSE)
}
\arguments{
\item{input_filename}{name of the input file,
as created by \code{\link{create_input_files_general}}}

\item{output_filename}{name of the output file}

\item{verbose}{shows output if TRUE}
}
\description{
Process an input file
}
\author{
Richel J.C. Bilderbeek
}
