% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_params_set.R
\name{create_params_set}
\alias{create_params_set}
\title{Create the set of parameters to use, for each of the
two experiment types}
\usage{
create_params_set(experiment_type)
}
\arguments{
\item{experiment_type}{type of experiment,
must be either 'general' or 'sampling'}
}
\description{
Create the set of parameters to use, for each of the
two experiment types
}
