% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pbd_params_file.R
\name{create_pbd_params_file}
\alias{create_pbd_params_file}
\title{Collect all PBD parameters of a \code{raket_werper} run and put these
in one \code{.csv} file called \code{results/pbd_params.csv}}
\usage{

  create_pbd_params_file(project_folder_name = get_raket_path("raket_werper"))
}
\arguments{
\item{project_folder_name}{name of project folder}
}
\value{
the name of the file created
}
\description{
Collect all PBD parameters of a \code{raket_werper} run and put these
in one \code{.csv} file called \code{results/pbd_params.csv}
}
