% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pbd_sim_out_file.R
\name{create_pbd_sim_out_file}
\alias{create_pbd_sim_out_file}
\title{Create the file to store the results of a PBD sim}
\usage{
create_pbd_sim_out_file(raket_params)
}
\arguments{
\item{raket_params}{\code{raket} parameters for one experiment,
as can be created by \code{\link{create_raket_params}}}
}
\value{
the filename
}
\description{
Create the file to store the results of a PBD sim
}
