% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_output_file.R
\name{create_posterior_files}
\alias{create_posterior_files}
\title{Process an input file}
\usage{
create_posterior_files(input_filename = input_filename,
  posterior_filesname = posterior_filesname, verbose = FALSE)
}
\arguments{
\item{input_filename}{name of the input file,
as created by \code{\link{create_input_files_general}}}

\item{posterior_filesname}{name of the output file}

\item{verbose}{shows output if TRUE}
}
\description{
Process an input file
}
\author{
Richel J.C. Bilderbeek
}
