% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sampling_params_set.R
\name{create_sampling_params_set}
\alias{create_sampling_params_set}
\title{Create the set of parameters for the experiment
to measure the effect of sampling}
\usage{
create_sampling_params_set(mcmc_chain_length = 1111000,
  sequence_length = 15000, n_replicates = 1, max_n_params = Inf)
}
\arguments{
\item{mcmc_chain_length}{length of the MCMC chain, in
number of states}

\item{sequence_length}{a DNA sequence length, in base pairs}

\item{n_replicates}{number of replicates per biological parameter set}

\item{max_n_params}{the maximum number of parameters created. Set to a lower
value in debugging}
}
\description{
Create the set of parameters for the experiment
to measure the effect of sampling
}
