% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table_n_extant_good_species.R
\name{create_table_n_extant_good_species}
\alias{create_table_n_extant_good_species}
\title{Create a}
\usage{
create_table_n_extant_good_species()
}
\description{
Create a
}
\author{
Richel J.C. Bilderbeek
}
