% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_test_pbd_params.R
\name{create_test_pbd_params}
\alias{create_test_pbd_params}
\title{Creates testing PBD params, using \link[becosys]{create_pbd_params}
with some parameters that are not that speciose}
\usage{
create_test_pbd_params(
  sirg = 0.1,
  siri = 0.15,
  scr = 0.2,
  erg = 0.01,
  eri = 0.02
)
}
\arguments{
\item{sirg}{speciation initiation rate of a good species}

\item{siri}{speciation initiation rate of an incipient species}

\item{scr}{speciation completion rate}

\item{erg}{extinction rate of a good species}

\item{eri}{extinction rate of an incipient species}
}
\description{
Creates testing PBD params, using \link[becosys]{create_pbd_params}
with some parameters that are not that speciose
}
\examples{
library(testthat)

pbd_params <- create_test_pbd_params()
expect_silent(becosys::check_pbd_params(pbd_params))
}
\author{
Richel J.C. Bilderbeek
}
