% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_test_pbd_paramses.R
\name{create_test_pbd_paramses}
\alias{create_test_pbd_paramses}
\title{Create testing PBD parameter sets}
\usage{
create_test_pbd_paramses()
}
\value{
a list of \code{pbd_params}
}
\description{
Create testing PBD parameter sets
}
