% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_test_raket_params.R
\name{create_test_raket_params}
\alias{create_test_raket_params}
\title{Create a test \code{raket_params}}
\usage{
create_test_raket_params(folder_name = peregrine::get_pff_tempdir())
}
\arguments{
\item{folder_name}{destination folder the parameters
will write to when run}
}
\value{
a test \code{raket_params},
  as can be created by \link{create_raket_params}
}
\description{
Create a test \code{raket_params}
}
\author{
Richel J.C. Bilderbeek
}
