% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_true_tree_file.R
\name{create_true_tree_file}
\alias{create_true_tree_file}
\title{Create the file for the true tree}
\usage{
create_true_tree_file(raket_params)
}
\arguments{
\item{raket_params}{\code{raket} parameters for one experiment,
as can be created by \code{\link{create_raket_params}}}
}
\description{
Create the file for the true tree
}
