% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raket_path.R
\name{get_raket_path}
\alias{get_raket_path}
\title{Get the full path of a file in the \code{inst/extdata} folder}
\usage{
get_raket_path(filename)
}
\arguments{
\item{filename}{name of the file}
}
\value{
the full path of the filename, if and only if
  the file is present. Will stop otherwise.
}
\description{
Get the full path of a file in the \code{inst/extdata} folder
}
\examples{
library(testthat)

expect_silent(get_raket_path("parameters.RDa"))
expect_error(get_raket_path("abs.ent"))
}
\author{
Richel J.C. Bilderbeek
}
