% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nltt_files_to_csv.R
\name{nltt_files_to_csv}
\alias{nltt_files_to_csv}
\title{Collect all nLTT values from \code{_nltts.RDa} files}
\usage{
nltt_files_to_csv(nltt_filenames = Sys.glob("nltt_*.RDa"), csv_filename)
}
\arguments{
\item{nltt_filenames}{names of the \code{_nltts.RDa} files,
as created by \code{create_nltt_file}}

\item{csv_filename}{name of the .csv file to be created}
}
\description{
Collect all nLTT values from \code{_nltts.RDa} files
}
\author{
Richel J.C. Bilderbeek
}
