% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbd_expected_n_extant.R
\name{pbd_expected_n_extant}
\alias{pbd_expected_n_extant}
\title{Get the median of the number of extant species from a PBD process.}
\usage{
pbd_expected_n_extant(crown_age, scr, sirg, siri, erg, eri, n_sims = 10)
}
\arguments{
\item{crown_age}{the crown age}

\item{scr}{speciation completion rate}

\item{sirg}{speciation initiation rate of a good species}

\item{siri}{speciation initiation rate of an incipient species}

\item{erg}{extinction rate of a good species}

\item{eri}{extinction rate of an incipient species}

\item{n_sims}{number of PBD simulations to take the median of. The
higher this number, the error on this calculation is decreased}
}
\description{
Get the median of the number of extant species from a PBD process.
}
\author{
Richel J.C. Bilderbeek
}
