% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbd_find_scenario.R
\name{pbd_find_scenario}
\alias{pbd_find_scenario}
\title{Create an example scenario}
\usage{
pbd_find_scenario(scr, sirg, siri, scenario, erg = 0, eri = 0,
  crown_age = 1, min_n_species = 1, max_n_species = 10000,
  min_n_subspecies = 2, max_n_subspecies = 10000, rng_seed = 42)
}
\arguments{
\item{scr}{Speciation Completion Rate}

\item{sirg}{Speciation Initiation Rate of Good Species}

\item{siri}{Speciation Initiation Rate of Incipient Species}

\item{sampling}{sampling method, can be 
'equal' (branch length of youngest equals oldest equals random)
'expected', 
'ylto' (youngest longer than oldest)
'rsty' (random shorter than youngest)
'rlto' (random longer than oldest)}
}
\description{
Create an example scenario
}
