% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbd_get_exp_n_lineages_pb.R
\name{pbd_get_exp_n_lineages_pb}
\alias{pbd_get_exp_n_lineages_pb}
\title{Get the expected number of good and incipient lineages from a protracted 
pure birth model}
\usage{
pbd_get_exp_n_lineages_pb(sirg, scr, siri, age, Ng, Ni)
}
\arguments{
\item{sirg}{The speciation initiation rate from a good species}

\item{scr}{The speciation completion rate}

\item{siri}{The speciation initiation rate from an incipient species}

\item{age}{The age}

\item{Ng}{The number of good species at t=0}

\item{Ni}{The number of incipient species at t=0}
}
\description{
Get the expected number of good and incipient lineages from a protracted 
pure birth model
}
\author{
Joris Damhuis
}
