% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbd_numspec_quantile_checked.R
\name{pbd_numspec_quantile_checked}
\alias{pbd_numspec_quantile_checked}
\title{Checked version of \code{\link[PBD]{pbd_numspec_quantile}}}
\usage{
pbd_numspec_quantile_checked(erg, eri, scr, sir, stem_age = NULL,
  crown_age = NULL, quantile)
}
\arguments{
\item{erg}{extinction rate of a good species}

\item{eri}{extinction rate of an incipient species}

\item{scr}{speciation completion rate}

\item{sir}{speciation initiation rate of both 
a good species and an incipient species}

\item{stem_age}{stem age. Set either the stem age or the crown age.}

\item{crown_age}{the crown age of the phylogeny}

\item{quantile}{the quantile, a value from, and including, zero to one}
}
\description{
Checked version of \code{\link[PBD]{pbd_numspec_quantile}}
}
\note{
uses equation 6 from \insertCite{etienne2014;textual}{raket}
}
\references{
\insertAllCited{}
}
