% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbd_sim_checked.R
\name{pbd_sim_checked}
\alias{pbd_sim_checked}
\title{Checked and safer version of PBD::pbd_sim}
\usage{
pbd_sim_checked(erg, eri, scr, sirg, siri, stem_age = NULL,
  crown_age = NULL, max_n_taxa = Inf)
}
\arguments{
\item{erg}{extinction rate of a good species}

\item{eri}{extinction rate of an incipient species}

\item{scr}{speciation completion rate}

\item{sirg}{speciation initiation rate of a good species}

\item{siri}{speciation initiation rate of an incipient species}

\item{stem_age}{stem age. Set either the stem age or the crown age.}

\item{crown_age}{crown age. Set either the stem age or the crown age.}

\item{max_n_taxa}{maximum number of taxa. If this value is exceeded,
the simulation is aborted and removed.}
}
\description{
Checked and safer version of PBD::pbd_sim
}
