% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkt_is_viable.R
\name{rkt_are_viable}
\alias{rkt_are_viable}
\title{Are these viable parameter combinations?}
\usage{
rkt_are_viable(crown_age, ergs, eris, scrs, sirgs, siris,
  quantile = 0.95)
}
\arguments{
\item{crown_age}{the crown age of the phylogeny}

\item{quantile}{the quantile, a value from, and including, zero to one}
}
\description{
Are these viable parameter combinations?
}
