% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkt_calc_bf.R
\name{rkt_calc_bf_bf}
\alias{rkt_calc_bf_bf}
\title{Calculate the Bayes factor from model A's point of view,
using the BayesFactor package}
\usage{
rkt_calc_bf_bf(log_likelihoods_a, log_likelihoods_b)
}
\arguments{
\item{log_likelihoods_a}{log-likelihoods of model A}

\item{log_likelihoods_b}{log-likelihoods of model B}
}
\value{
the Bayes factor from model A's point of view
}
\description{
Calculate the Bayes factor from model A's point of view,
using the BayesFactor package
}
\references{
* [1] Drummond, Alexei J., and Remco R. Bouckaert. Bayesian evolutionary analysis with BEAST. Cambridge University Press, 2015. \cr
   * [2] Newton, Michael A., and Adrian E. Raftery. "Approximate Bayesian inference with the weighted likelihood bootstrap." Journal of the Royal Statistical Society. Series B (Methodological) (1994): 3-48. \cr
}
\author{
Richel J.C. Bilderbeek
}
