% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkt_calc_bf.R
\name{rkt_calc_bf_rse}
\alias{rkt_calc_bf_rse}
\title{Calculate the Bayes factor from model A's point of view}
\usage{
rkt_calc_bf_rse(log_likelihoods_a, log_likelihoods_b)
}
\arguments{
\item{log_likelihoods_a}{log-likelihoods of model A}

\item{log_likelihoods_b}{log-likelihoods of model B}
}
\value{
the Bayes factor from model A's point of view
}
\description{
Calculate the Bayes factor from model A's point of view
}
\author{
Rampal S. Etienne, modified by Richel J.C. Bilderbeek
}
