% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkt_get_clock_models.R
\name{rkt_get_clock_models}
\alias{rkt_get_clock_models}
\title{Get the clock models used in this research}
\usage{
rkt_get_clock_models()
}
\description{
Get the clock models used in this research
}
