% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkt_get_experiment_types.R
\name{rkt_get_experiment_types}
\alias{rkt_get_experiment_types}
\title{Get the two types of experiments, which
are 'general' (to produce a general data set),
and 'sampling' (to investigate the worst-case
effect of sampling)}
\usage{
rkt_get_experiment_types()
}
\description{
Get the two types of experiments, which
are 'general' (to produce a general data set),
and 'sampling' (to investigate the worst-case
effect of sampling)
}
