% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkt_get_extinction_rates.R
\name{rkt_get_ext_rates}
\alias{rkt_get_ext_rates}
\title{Get the extinction rates used in this research}
\usage{
rkt_get_ext_rates()
}
\description{
Get the extinction rates used in this research
}
