% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkt_get_max_n_cols.R
\name{rkt_get_max_n_cols}
\alias{rkt_get_max_n_cols}
\title{Get the number of columns each dataset must maximally have.
Together with \code{\link{rkt_get_max_n_rows}}, the size
of a data frame is determined. A data frame of such size is
tested to fit in memory twice.}
\usage{
rkt_get_max_n_cols()
}
\description{
Get the number of columns each dataset must maximally have.
Together with \code{\link{rkt_get_max_n_rows}}, the size
of a data frame is determined. A data frame of such size is
tested to fit in memory twice.
}
