% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkt_get_max_n_rows.R
\name{rkt_get_max_n_rows}
\alias{rkt_get_max_n_rows}
\title{Get the number of rows each dataset must maximally have.
Together with \code{\link{rkt_get_max_n_cols}}, the size
of a data frame is determined. A data frame of such size is
tested to fit in memory twice.}
\usage{
rkt_get_max_n_rows()
}
\description{
Get the number of rows each dataset must maximally have.
Together with \code{\link{rkt_get_max_n_cols}}, the size
of a data frame is determined. A data frame of such size is
tested to fit in memory twice.
}
