% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkt_get_n_param_combos.R
\name{rkt_get_n_param_combos}
\alias{rkt_get_n_param_combos}
\title{Get the number of parameter combinations}
\usage{
rkt_get_n_param_combos()
}
\description{
Get the number of parameter combinations
}
