% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkt_get_n_params.R
\name{rkt_get_n_params}
\alias{rkt_get_n_params}
\title{Get the number of parameters used in this research,
for example, speciation completion rate, speciation
initiation rate of a good species, etc.}
\usage{
rkt_get_n_params()
}
\description{
Get the number of parameters used in this research,
for example, speciation completion rate, speciation
initiation rate of a good species, etc.
}
