% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkt_get_n_replicates.R
\name{rkt_get_n_replicates}
\alias{rkt_get_n_replicates}
\title{Get the number of replicates per parameter combination.}
\usage{
rkt_get_n_replicates()
}
\description{
Get the number of replicates per parameter combination.
}
