% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkt_get_sampling_methods.R
\name{rkt_get_sampling_methods}
\alias{rkt_get_sampling_methods}
\title{Get the sampling methods used}
\usage{
rkt_get_sampling_methods()
}
\description{
Get the sampling methods used
}
\author{
Richel J.C. Bilderbeek
}
