% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkt_get_spec_init_rates.R
\name{rkt_get_spec_init_rates}
\alias{rkt_get_spec_init_rates}
\title{Get the speciation initiation rates used in this research
Old version: { 0.1, 0.5, 1.0 }}
\usage{
rkt_get_spec_init_rates()
}
\description{
Get the speciation initiation rates used in this research
Old version: { 0.1, 0.5, 1.0 }
}
