% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkt_is_viable.R
\name{rkt_is_viable}
\alias{rkt_is_viable}
\title{Is this a viable parameter combination?}
\usage{
rkt_is_viable(erg, eri, sirg, siri)
}
\arguments{
\item{erg}{extinction rate of a good species}

\item{eri}{extinction rate of an incipient species}

\item{sirg}{speciation initiation rate of a good species}

\item{siri}{speciation initiation rate of an incipient species}
}
\description{
Is this a viable parameter combination?
}
