% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkt_plot.R
\name{rkt_plot}
\alias{rkt_plot}
\title{Plot violin plot}
\usage{
rkt_plot(df_long = utils::read.csv("result_long.csv"),
  pdf_filename = tempfile(fileext = ".pdf"))
}
\arguments{
\item{df_long}{data frame in the long form}
}
\description{
Plot violin plot
}
