% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkt_run_from_file.R
\name{rkt_run_from_file}
\alias{rkt_run_from_file}
\title{Run a 'raket' experiment from a filename}
\usage{
rkt_run_from_file(parameters_filename)
}
\arguments{
\item{parameters_filename}{name of a file that contains
a \code{raket_params},
as can be created by \link{create_raket_params}}
}
\description{
Run a 'raket' experiment from a filename
}
\author{
Richel J.C. Bilderbeek
}
