% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_raket_from_file.R
\name{run_raket_from_file}
\alias{run_raket_from_file}
\title{Run a raket experiment from a filename}
\usage{
run_raket_from_file(parameters_filename)
}
\arguments{
\item{parameters_filename}{name of a file that contains
a \code{raket_params},
as can be created by \link{create_params_raket}}
}
\description{
Run a raket experiment from a filename
}
\author{
Richel J.C. Bilderbee
}
