% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_median_nltts.R
\name{calc_median_nltts}
\alias{calc_median_nltts}
\title{For each input file, calculate the median nLTT statistic.
Put these, with the parameters in a data frame}
\usage{
calc_median_nltts(input_filenames, output_filename)
}
\description{
For each input file, calculate the median nLTT statistic.
Put these, with the parameters in a data frame
}
\author{
Richel J.C. Bilderbeek
}
