% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_raket_paramses.R
\name{check_raket_paramses}
\alias{check_raket_paramses}
\title{Check if the argument is a list of \code{raket_params}}
\usage{
check_raket_paramses(raket_paramses)
}
\description{
Check if the argument is a list of \code{raket_params}
}
\author{
Richel J.C. Bilderbeek
}
