% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_general_params_set.R
\name{create_general_params_set}
\alias{create_general_params_set}
\title{Creates the parameter set for the general exploration}
\usage{
create_general_params_set(mcmc = beautier::create_mcmc(chain_length = 1111000,
  store_every = 1000), sequence_length = 15000, n_replicates = 1)
}
\arguments{
\item{mcmc}{MCMC options,
as created by \link[beautier]{create_mcmc}}

\item{sequence_length}{DNA alignment sequence length,
in number of nucleotides}

\item{n_replicates}{number of replicates per biological parameter set}
}
\description{
Creates the parameter set for the general exploration
}
\author{
Richel Bilderbeek
}
