% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_general_params_set.R
\name{create_general_params_set}
\alias{create_general_params_set}
\title{Creates the parameter set for the general exploration}
\usage{
create_general_params_set(project_folder_name = getwd(),
  crown_age = 15, max_n_params = Inf, mcmc_chain_length = 1111000,
  n_replicates = 1, sequence_length = 15000)
}
\arguments{
\item{project_folder_name}{name of project folder}

\item{crown_age}{the crown age of the phylogeny}

\item{max_n_params}{the maximum number of parameters created. Set to a lower
value in debugging}

\item{mcmc_chain_length}{length of the MCMC chain, in
number of states}

\item{n_replicates}{number of replicates per biological parameter set}

\item{sequence_length}{a DNA sequence length, in base pairs}
}
\value{
a list of which each element is a \code{raket_params}
}
\description{
Creates the parameter set for the general exploration
}
\author{
Richel J.C. Bilderbeek
}
