% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_input_files.R
\name{create_input_files}
\alias{create_input_files}
\title{Creates the parameter files in the article
Now every parameter setting has its own seed}
\usage{
create_input_files(mcmc_length = 1e+06)
}
\arguments{
\item{mcmc_length}{number of states in the MCMC chain}

\item{minimal_ess}{minimal posterior Effect Sample Size}
}
\value{
The filenames of all parameter files created
}
\description{
Creates the parameter files in the article
Now every parameter setting has its own seed
}
\author{
Richel Bilderbeek
}
