% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_input_files.R
\name{create_input_files_general}
\alias{create_input_files_general}
\title{Creates the parameter files in the article for the general exploration}
\usage{
create_input_files_general(general_params_set = create_general_params_set(),
  folder_name = getwd())
}
\arguments{
\item{folder_name}{name of the folder where all files are created}

\item{mcmc}{MCMC options, as created by \link[beautier]{create_mcmc}}

\item{sequence_length}{DNA alignment sequence length,
in number of nucleotides}

\item{n_replicates}{number of replicates per biological parameter set}
}
\value{
The filenames of all parameter files created
}
\description{
Creates the parameter files in the article for the general exploration
}
\author{
Richel Bilderbeek
}
