% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_input_files.R
\name{create_input_files_general}
\alias{create_input_files_general}
\alias{create_files_raket_paramses}
\title{Creates the parameter files in the article for the general exploration}
\usage{
create_input_files_general(raket_paramses = create_general_params_set())
}
\arguments{
\item{general_params_set}{the set of parameters,
as created by \code{\link{create_general_params_set}}}

\item{project_folder_name}{name of the project folder.
The name of this folder must be \code{raket_werper}.}
}
\value{
The filenames of all parameter files created
}
\description{
Creates the parameter files in the article for the general exploration
}
\author{
Richel Bilderbeek
}
