% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_params_set.R
\name{create_params_set}
\alias{create_params_set}
\title{Create the set of parameters to use, for each of the
two experiment types}
\usage{
create_params_set(experiment_type, mcmc = beautier::create_mcmc(chain_length =
  1111000, store_every = 1000), sequence_length = 15000, n_replicates = 1,
  max_n_params = Inf)
}
\arguments{
\item{experiment_type}{type of experiment,
must be either 'general' or 'sampling'}

\item{mcmc}{MCMC options,
as created by \link[beautier]{create_mcmc}}

\item{sequence_length}{DNA alignment sequence length,
in number of nucleotides}

\item{n_replicates}{number of replicates per biological parameter set}

\item{max_n_params}{the maximum number of parameters created. Set to a lower
value in debugging}
}
\value{
a list of parameters
}
\description{
Create the set of parameters to use, for each of the
two experiment types
}
