% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_raket_params.R
\name{create_raket_params}
\alias{create_raket_params}
\title{Create the parameters for one experiment.
Run one point of the experiment}
\usage{
create_raket_params(pbd_params,
  twinning_params = pirouette::create_twinning_params(),
  alignment_params, gen_model_select_params, crown_age,
  crown_age_sigma = 0.01, sampling_method, mcmc, tree_sim_rng_seed,
  beast2_rng_seed, site_model = "JC69", clock_model = "strict")
}
\arguments{
\item{pbd_params}{a PBD parameter set,
as returned by \code{\link[becosys]{create_pbd_params}}}

\item{twinning_params}{parameters for creating a twin tree,
as can be created by \code{\link[pirouette]{create_twinning_params}}}

\item{alignment_params}{parameters for creating an alignment,
as can be created by \code{\link[pirouette]{create_alignment_params}}}

\item{gen_model_select_params}{a generative model selection parameter,
which will select the generative model to be run in
the Bayesian inference.
Use \code{\link[pirouette]{create_gen_model_select_params}} to create
this parameter set}

\item{crown_age}{crown age, in time units}

\item{crown_age_sigma}{the standard deviation of MRCA prior's
distribution, in time units}

\item{sampling_method}{method how the incipient species representing a
species is picked. Must be in \link{rkt_get_sampling_methods}.}

\item{mcmc}{MCMC options, as created by \link[beautier]{create_mcmc}}

\item{tree_sim_rng_seed}{RNG seed used in the creation of the true
phylogenetic tree}

\item{beast2_rng_seed}{RNG seed in the creation of the BEAST2 posterior
(based on the alignment)}

\item{site_model}{site model assumed in the BEAST2 inference}

\item{clock_model}{clock model assumed in the BEAST2 inference}
}
\description{
Create the parameters for one experiment.
Run one point of the experiment
}
\author{
Richel J.C. Bilderbeek
}
