% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_raket_params.R
\name{create_raket_params}
\alias{create_raket_params}
\title{Create the parameters for one experiment.
Run one point of the experiment}
\usage{
create_raket_params(sirg, siri, scr, erg, eri, crown_age,
  crown_age_sigma = 0.01, sampling_method, mutation_rate,
  sequence_length, mcmc, tree_sim_rng_seed, alignment_rng_seed,
  beast2_rng_seed, site_model = "JC69", clock_model = "strict")
}
\arguments{
\item{sirg}{speciation rate of a good species,
per lineage probability per time unit}

\item{siri}{speciation rate of an incipient species,
per lineage probability per time unit}

\item{scr}{speciation completion rate (of
incipient species),
per lineage probability per time unit}

\item{erg}{extinction rate of a good species,
per lineage probability per time unit}

\item{eri}{extinction rate of an incipient species,
per lineage probability per time unit}

\item{crown_age}{crown age, in time units}

\item{crown_age_sigma}{the standard deviation of MRCA prior's
distribution, in time units}

\item{sampling_method}{method how the incipient species representing a
species is picked. Must be in \link{rkt_get_sampling_methods}.}

\item{mutation_rate}{per-nucleotide probability to mutate per time unit}

\item{sequence_length}{length of simulated DNA alignment,
number of base pairs}

\item{mcmc}{MCMC options, as created by \link[beautier]{create_mcmc}}

\item{tree_sim_rng_seed}{RNG seed used in the creation of the true
phylogenetic tree}

\item{alignment_rng_seed}{RNG seed used in the creation of the simulated
alignment (from the true phylogenetic tree)}

\item{beast2_rng_seed}{RNG seed in the creation of the BEAST2 posterior
(based on the alignment)}

\item{site_model}{site model assumed in the BEAST2 inference}

\item{clock_model}{clock model assumed in the BEAST2 inference}
}
\description{
Create the parameters for one experiment.
Run one point of the experiment
}
\author{
Richel J.C. Bilderbeek
}
