% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_raket_params.R
\name{create_raket_params}
\alias{create_raket_params}
\title{Create the parameters for one experiment.
Run one point of the experiment}
\usage{
create_raket_params(pbd_params,
  twinning_params = pirouette::create_twinning_params(),
  alignment_params, gen_model_select_params, best_model_select_params,
  inference_params, sampling_method)
}
\arguments{
\item{pbd_params}{a PBD parameter set,
as returned by \code{\link[becosys]{create_pbd_params}}}

\item{twinning_params}{parameters for creating a twin tree,
as can be created by \code{\link[pirouette]{create_twinning_params}}}

\item{alignment_params}{parameters for creating an alignment,
as can be created by \code{\link[pirouette]{create_alignment_params}}}

\item{gen_model_select_params}{a generative model selection parameter,
which will select the generative model to be run in
the Bayesian inference.
Use \link[pirouette]{create_gen_model_select_param} to create
this parameter set}

\item{best_model_select_params}{a parameter set to specify to
pick the inference model with the highest evidene (also:
marginal likelihood),
as can be created by
\code{\link[pirouette]{create_best_model_select_param}}}

\item{inference_params}{the parameters used in all Bayesian inference,
as created by \link[pirouette]{create_inference_params}}

\item{sampling_method}{method how the incipient species representing a
species is picked. Must be in \link{rkt_get_sampling_methods}.}
}
\description{
Create the parameters for one experiment.
Run one point of the experiment
}
\author{
Richel J.C. Bilderbeek
}
