% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_raket_params.R
\name{create_raket_params}
\alias{create_raket_params}
\alias{create_params_raket}
\title{Create the parameters for one experiment.
Run one point of the experiment}
\usage{
create_raket_params(
  pbd_params,
  pir_params,
  sampling_method,
  true_tree_filename = file.path(dirname(pir_params$alignment_params$fasta_filename),
    "pbd.newick"),
  pbd_sim_out_filename = file.path(dirname(pir_params$alignment_params$fasta_filename),
    "pbd_sim_out.RDa")
)
}
\arguments{
\item{pbd_params}{a PBD parameter set,
as returned by \code{\link[becosys]{create_pbd_params}}}

\item{pir_params}{\code{pirouette} parameter set,
as created by \link[pirouette]{create_pir_params}}

\item{sampling_method}{method how the incipient species representing a
species is picked. Must be in \link{rkt_get_sampling_methods}.}

\item{true_tree_filename}{path to the file to store the true
phylogentic tree to. Tree will be stored in Newick format.}

\item{pbd_sim_out_filename}{path to file to save the PBD simulation's
output to. Data will be stored as R Data (\code{.RDa}).}
}
\description{
Create the parameters for one experiment.
Run one point of the experiment
}
\author{
Richel J.C. Bilderbeek
}
