% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sampling_params_set.R
\name{create_sampling_params_set}
\alias{create_sampling_params_set}
\title{Create the set of parameters for the experiment
to measure the effect of sampling}
\usage{
create_sampling_params_set(
  project_folder_name = file.path(peregrine::get_pff_tempdir(), "raket_werper"),
  n_replicates = 1
)
}
\arguments{
\item{project_folder_name}{name of project folder}

\item{n_replicates}{number of replicates per biological parameter set}
}
\description{
Create the set of parameters for the experiment
to measure the effect of sampling
}
