% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_test_params_set.R
\name{create_test_params_set}
\alias{create_test_params_set}
\title{Create the parameter set used in testing}
\usage{
create_test_params_set()
}
\value{
a list of which each element is a \code{raket_params}
}
\description{
Create the parameter set used in testing
}
\author{
Richel J.C. Bilderbeek
}
