% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_param_doc.R
\name{default_params_doc}
\alias{default_params_doc}
\title{This function does nothing. It is intended to inherit is parameters'
documentation.}
\usage{
default_params_doc(crown_age, experiment_type, max_n_params, mcmc, n_replicates,
  sequence_length, verbose)
}
\arguments{
\item{crown_age}{the crown age of the phylogeny}

\item{experiment_type}{type of experiment,
must a member of \code{\link{rkt_get_experiment_types}}}

\item{max_n_params}{the maximum number of parameters created. Set to a lower
value in debugging}

\item{mcmc}{one mcmc object,
as returned by \code{\link[beautier]{create_mcmc}}}

\item{n_replicates}{number of replicates per biological parameter set}

\item{sequence_length}{a DNA sequence length, in base pairs}

\item{verbose}{if TRUE, additional information is displayed, that
is potentially useful in debugging}
}
\description{
This function does nothing. It is intended to inherit is parameters'
documentation.
}
\note{
This is an internal function, so it should be marked with
  \code{@noRd}. This is not done, as this will disallow all
  functions to find the documentation parameters
}
\author{
Richel J.C. Bilderbeek
}
