% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_param_doc.R
\name{default_params_doc}
\alias{default_params_doc}
\title{This function does nothing. It is intended to inherit is parameters'
documentation.}
\usage{
default_params_doc(alignment_params, best_model_select_params, crown_age,
  erg, ergs, eri, eris, experiment_type, gen_model_select_params,
  inference_params, max_n_params, mcmc, mcmc_chain_length, n_replicates,
  pbd_params, pir_params, quantile, raket_params, sampling_method, scr,
  scrs, sequence_length, sir, sirg, sirgs, siri, siris, stem_age,
  twinning_params, verbose)
}
\arguments{
\item{alignment_params}{parameters for creating an alignment,
as can be created by \code{\link[pirouette]{create_alignment_params}}}

\item{crown_age}{the crown age of the phylogeny}

\item{erg}{extinction rate of a good species}

\item{ergs}{extinction rates for good species}

\item{eri}{extinction rate of an incipient species}

\item{eris}{extinction rates for incipient species}

\item{experiment_type}{type of experiment,
must a member of \code{\link{rkt_get_experiment_types}}}

\item{max_n_params}{the maximum number of parameters created. Set to a lower
value in debugging}

\item{mcmc}{one mcmc object,
as returned by \code{\link[beautier]{create_mcmc}}}

\item{mcmc_chain_length}{length of the MCMC chain, in
number of states}

\item{n_replicates}{number of replicates per biological parameter set}

\item{pbd_params}{a PBD parameter set,
as returned by \code{\link[becosys]{create_pbd_params}}}

\item{pir_params}{\code{pirouette} parameter set,
as created by \link[pirouette]{create_pir_params}}

\item{quantile}{the quantile, a value from, and including, zero to one}

\item{raket_params}{\code{raket} parameters for one experiment,
as can be created by \code{\link{create_raket_params}}}

\item{sampling_method}{method how the incipient species representing a
species is picked. Must be in \link{rkt_get_sampling_methods}.}

\item{scr}{speciation completion rate}

\item{scrs}{speciation completion rates}

\item{sequence_length}{a DNA sequence length, in base pairs}

\item{sir}{speciation initiation rate of both
a good species and an incipient species}

\item{sirg}{speciation initiation rate of a good species}

\item{sirgs}{speciation initiation rates of good species}

\item{siri}{speciation initiation rate of an incipient species}

\item{siris}{speciation initiation rates of incipient species}

\item{stem_age}{stem age. Set either the stem age or the crown age.}

\item{twinning_params}{parameters for creating a twin tree,
as can be created by \code{\link[pirouette]{create_twinning_params}}}

\item{verbose}{if TRUE, additional information is displayed, that
is potentially useful in debugging}
}
\description{
This function does nothing. It is intended to inherit is parameters'
documentation.
}
\note{
This is an internal function, so it should be marked with
  \code{@noRd}. This is not done, as this will disallow all
  functions to find the documentation parameters
}
\author{
Richel J.C. Bilderbeek
}
