% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_param_doc.R
\name{default_params_doc}
\alias{default_params_doc}
\title{This function does nothing. It is intended to inherit is parameters'
documentation.}
\usage{
default_params_doc(crown_age, erg, eri, experiment_type, max_n_params,
  mcmc, mcmc_chain_length, n_replicates, pbd_params, quantile,
  raket_params, scr, sequence_length, sir, sirg, siri, stem_age,
  twinning_params, verbose)
}
\arguments{
\item{crown_age}{the crown age of the phylogeny}

\item{erg}{extinction rate of a good species}

\item{eri}{extinction rate of an incipient species}

\item{experiment_type}{type of experiment,
must a member of \code{\link{rkt_get_experiment_types}}}

\item{max_n_params}{the maximum number of parameters created. Set to a lower
value in debugging}

\item{mcmc}{one mcmc object,
as returned by \code{\link[beautier]{create_mcmc}}}

\item{mcmc_chain_length}{length of the MCMC chain, in
number of states}

\item{n_replicates}{number of replicates per biological parameter set}

\item{pbd_params}{a PBD parameter set,
as returned by \code{\link[becosys]{create_pbd_params}}}

\item{quantile}{the quantile, a value from, and including, zero to one}

\item{raket_params}{\code{raket} parameters for one experiment,
as can be created by \code{\link{create_raket_params}}}

\item{scr}{speciation completion rate}

\item{sequence_length}{a DNA sequence length, in base pairs}

\item{sir}{speciation initiation rate of both
a good species and an incipient species}

\item{sirg}{speciation initiation rate of a good species}

\item{siri}{speciation initiation rate of an incipient species}

\item{stem_age}{stem age. Set either the stem age or the crown age.}

\item{twinning_params}{parameters for creating a twin tree,
as can be created by \code{\link[pirouette]{create_twinning_params}}}

\item{verbose}{if TRUE, additional information is displayed, that
is potentially useful in debugging}
}
\description{
This function does nothing. It is intended to inherit is parameters'
documentation.
}
\note{
This is an internal function, so it should be marked with
  \code{@noRd}. This is not done, as this will disallow all
  functions to find the documentation parameters
}
\author{
Richel J.C. Bilderbeek
}
