% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkt_create_data_frame.R
\name{rkt_create_data_frame}
\alias{rkt_create_data_frame}
\title{Create a skeleton data frame for each data set}
\usage{
rkt_create_data_frame(n_replicates, n_nltts, experiment_type,
  max_n_params = Inf)
}
\arguments{
\item{n_replicates}{number of replicates}

\item{n_nltts}{number of nLTT statistics per file}

\item{experiment_type}{type of experiment,
must be either 'general' or 'sampling'}

\item{max_n_params}{the maximum number of parameters created. Set to a lower
value in debugging}
}
\description{
Create a skeleton data frame for each data set
}
