% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkt_create_data_frame.R
\name{rkt_create_data_frame}
\alias{rkt_create_data_frame}
\title{Create a skeleton data frame for each data set}
\usage{
rkt_create_data_frame(n_files, n_nltts, experiment_type)
}
\arguments{
\item{n_files}{number of parameter files}

\item{n_nltts}{number of nLTT statistics per file}

\item{experiment_type}{type of experiment,
must be either 'general' or 'sampling'}
}
\description{
Create a skeleton data frame for each data set
}
