% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkt_run.R
\name{rkt_run}
\alias{rkt_run}
\title{Run one point of the experiment}
\usage{
rkt_run(parameters,
  beast_jar_path = beastier::get_default_beast2_jar_path(),
  verbose = FALSE)
}
\arguments{
\item{parameters}{biological and methodological parameters,
as created by \code{\link{create_params}}}

\item{beast_jar_path}{the path of beast.jar}

\item{verbose}{set to TRUE for more (debug) output}
}
\description{
Run one point of the experiment
}
\author{
Richel J.C. Bilderbeek
}
