% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkt_run.R
\name{rkt_run}
\alias{rkt_run}
\title{Run one point of the experiment}
\usage{
rkt_run(parameters, verbose = FALSE)
}
\arguments{
\item{parameters}{biological and methodological parameters,
as created by \code{\link{create_params}}}

\item{verbose}{set to TRUE for more (debug) output}

\item{beast2_path}{the path of either
the BEAST2 binary (\code{beast})}
}
\description{
Run one point of the experiment
}
\author{
Richel J.C. Bilderbeek
}
