% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkt_run.R
\name{rkt_run}
\alias{rkt_run}
\title{Run one point of the experiment}
\usage{
rkt_run(raket_params)
}
\arguments{
\item{raket_params}{\code{raket} parameters for one experiment,
as can be created by \code{\link{create_raket_params}}}
}
\description{
Run one point of the experiment
}
\author{
Richel J.C. Bilderbeek
}
